feat!: base `MotionController` on `TrackController`
BREAKING CHANGE: Like `AnimationController`, `MotionController` used to synchronously emit
its initial value when an animation was started. This is no longer the case.

This can subtly change your state machines or animation timings.
